Added new work machine to setup and fixed layouts
Using a weird symlink based modular layout setup as the layoutHooks type is a mess
for xmonad.  This creates a layout import that has a changing type (so there are
no type conflict for the different modified versions).  The setup.sh script will
try to link the correct setup based on hostname, if one doesn't exist, uses the
default one.  This should probably copy the default into a new machine specific one
to allow modification, but that needs to be ironed out first.