Added MPD support into system settings
Some of the configs for creating the dzen bars will probably need some tweaking
and the conky config could use some love for MPD.  Using a `Bool` for the
setting in each of the machine specific config files.  Also added some changes
for the Laurie layout set (need to do this for all of the machines).